requirements.txt
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
curl_cffi==0.6.2
duckduckgo_search==5.3.0
exceptiongroup==1.2.1
Flask==2.3.3
flask-mysql-connector==1.1.0
Flask-WTF==1.2.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysql-connector-python==9.0.0
numpy==1.26.4
ollama==0.1.8
orjson==3.10.1
packaging==24.0
pandas==2.2.2
pgvector==0.2.5
phidata==2.4.20
pillow==10.3.0
pluggy==1.5.0
protobuf==4.25.3
psycopg==3.1.18
psycopg-binary==3.1.18
pyarrow==15.0.2
pycparser==2.22
pydantic==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
pydeck==0.8.1b0
Pygments==2.17.2
pypdf==4.2.0
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
streamlit==1.33.0
tenacity==8.2.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
watchdog==4.0.1
Werkzeug==3.0.3
WTForms==3.1.2
